requirements.txt
Django==1.5.1
South==0.8.2
django-grappelli==2.4.5
django-guardian==1.1.1
docutils==0.11
psycopg2==2.5.1
wsgiref==0.1.2
